+++ src/datawizard/malloc.c
@@ -315,7 +315,7 @@ end:
 #endif
 	if (ret == 0)
 	{
-		STARPU_ASSERT_MSG(*A, "Failed to allocated memory of size %ld b\n", dim);
+		STARPU_ASSERT_MSG(*A, "Failed to allocated memory of size %ld b\n", (unsigned long)dim);
 	}
 	else if (flags & STARPU_MALLOC_COUNT)
 	{
@@ -1037,7 +1037,7 @@ starpu_free_on_node_flags(unsigned dst_node, uintptr_t addr, size_t size, int fl
 
 void starpu_malloc_on_node_set_default_flags(unsigned node, int flags)
 {
-	STARPU_ASSERT_MSG(node < STARPU_MAXNODES, "bogus node value %d given to starpu_malloc_on_node_set_default_flags\n", node);
+	STARPU_ASSERT_MSG(node < STARPU_MAXNODES, "bogus node value %u given to starpu_malloc_on_node_set_default_flags\n", node);
 	malloc_on_node_default_flags[node] = flags;
 }
